Add support for building on clang 19

+++ common/nvmeib_larray.h
@@ -22,7 +22,7 @@
 		memset((_a).a, 0, sizeof((_a).a)); \
 		BUILD_BUG_ON(sizeof((_a).a[0]) < sizeof(int));\
 		for (LACONCAT(i, __LINE__) = 0; \
-				LACONCAT(i, __LINE__) < ARRAY_SIZE((_a).a) - 1; \
+				LACONCAT(i, __LINE__) < (int)ARRAY_SIZE((_a).a) - 1; \
 				++LACONCAT(i, __LINE__)) \
 			*((int *)&((_a).a[LACONCAT(i, __LINE__)])) = \
 				LACONCAT(i, __LINE__) + 1; \
